Changed the FindBook class to a FilterBook class Modified DeleteBook slightly by taking out any unnecessary classes

+++ src/main/java/com/azure/app/OptionChecker.java
@@ -1,9 +1,19 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 package com.azure.app;
 
 import java.io.File;
 
 public class OptionChecker {
 
+    /**
+     * Checks the user's string and makes sure its valid.
+     *
+     * @param input - String the user enteredd
+     * @return -  boolean : true - the String is valid
+     * false - the String isn't valid
+     */
     public boolean validateString(String input) {
         if (input.isEmpty()) {
             System.out.println("Please enter a value: ");
@@ -18,6 +28,14 @@ public class OptionChecker {
         return !onlySpaces;
     }
 
+    /**
+     * Verifies the image path to make sure its valid and right
+     * Checks the extension to make sure its png., jpg., gif.
+     *
+     * @param image - File with the supposed image
+     * @return - boolean : true - if image path is correct
+     *                     false - otherwise.
+     */
     public boolean checkImage(File image) {
         if (!image.isFile()) {
             System.out.println("Please write out a valid image file path.");
@@ -35,6 +53,12 @@ public class OptionChecker {
         return false;
     }
 
+    /**
+     * Verifies the user's input to make sure its either y/n
+     * @param choice - String containing a choice
+     * @return - boolean : true - the String is either a y/n
+     *                     false - the String is neither
+     */
     public boolean checkYesOrNo(String choice) {
         if (!choice.contentEquals("Y") && !choice.contentEquals("y")
             && !choice.contentEquals("N") && !choice.contentEquals("n")) {
@@ -45,6 +69,12 @@ public class OptionChecker {
         }
     }
 
+    /**
+     * Verifies the author's name to make sure there's a first and last name.
+     * @param author - String with the author's name
+     * @return boolean : true - if String is valid
+     *                   false - if String isn't valid
+     */
     public boolean validateAuthor(String[] author) {
         if (author.length == 0) {
             System.out.println("Please enter their name.");
@@ -57,6 +87,12 @@ public class OptionChecker {
         return true;
     }
 
+    /**
+     * Checks the file to see if its contents match the book and that it's an actual json file.
+     * @param f - File that needs to be checked
+     * @param b - the Book object that it's being compared to
+     * @return
+     */
     public boolean checkFile(File f, Book b) {
         if (f == null) {
             return false;
@@ -66,8 +102,13 @@ public class OptionChecker {
             && f.getAbsolutePath().endsWith(".json");
     }
 
-
-    public int checkOption(String option, int max) {
+    /**
+     * Checks the String option and parses it to an integer if it fits within the desired range or is actually a numeral
+     * @param option - String that the user picked
+     * @param max - a long containing the maximum number of options
+     * @return - valid integer option, or else -1 signifies that the choice was invalid
+     */
+    public int checkOption(String option, long max) {
         if (option.equalsIgnoreCase("q")) {
             return 0;
         }
